Solucionado el issue #6 , ahora los cambios en el DigestMethod del objeto XMLAdvanceSignature aplican también al SigningCertificate

+++ src/main/java/es/uji/crypto/xades/jxades/util/Base64.java
@@ -1246,7 +1246,7 @@ public class Base64
         boolean dontGunzip = (options & DONT_GUNZIP) != 0;
         if( (bytes != null) && (bytes.length >= 4) && (!dontGunzip) ) {
             
-            int head = ((int)bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
+            int head = (bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
             if( java.util.zip.GZIPInputStream.GZIP_MAGIC == head )  {
                 java.io.ByteArrayInputStream  bais = null;
                 java.util.zip.GZIPInputStream gzis = null;
